+++ CHANGELOG.md
@@ -1,3 +1,41 @@
+# 1.2.0 (Wed Nov 26 2025)
+
+#### 🚀 Enhancement
+
+- 276 Feature Implement Quality Control Framework [#277](https://github.com/sensein/senselab/pull/277) ([@ibevers](https://github.com/ibevers) rfbrito@node1703.inband)
+- CI and Feature Extraction Optimization, nemo_toolkit containerization, pyannote-audio bump, and more minor edits [#407](https://github.com/sensein/senselab/pull/407) ([@fabiocat93](https://github.com/fabiocat93))
+
+#### 🐛 Bug Fix
+
+- Merge Rahul's QC changes [#365](https://github.com/sensein/senselab/pull/365) (rfbrito@node1703.inband [@ibevers](https://github.com/ibevers) [@Rahul-Brito](https://github.com/Rahul-Brito))
+- 298 Task Update Bioacoustic QC Taxonomy Tree with B2AI Activities [#348](https://github.com/sensein/senselab/pull/348) ([@ibevers](https://github.com/ibevers))
+- 347 Task Add Snorkel Automated Review [#356](https://github.com/sensein/senselab/pull/356) ([@ibevers](https://github.com/ibevers))
+- Update Quality Control Metrics based on Satra's Feedback [#357](https://github.com/sensein/senselab/pull/357) ([@satra](https://github.com/satra) [@ibevers](https://github.com/ibevers) [@fabiocat93](https://github.com/fabiocat93))
+- 296 Task Recursive Bioacoustic QC Class Structure [#345](https://github.com/sensein/senselab/pull/345) ([@ibevers](https://github.com/ibevers))
+- 326 Task Calculate Evaluations over Windows [#344](https://github.com/sensein/senselab/pull/344) ([@ibevers](https://github.com/ibevers))
+- 297 Task Parallelize Bioacoustic QC [#322](https://github.com/sensein/senselab/pull/322) ([@ibevers](https://github.com/ibevers))
+- 300 task add bioacoustic qc checks based on b2ai [#320](https://github.com/sensein/senselab/pull/320) ([@ibevers](https://github.com/ibevers))
+- bioacoustic qc store filepath if provided with waveform [#336](https://github.com/sensein/senselab/pull/336) ([@ibevers](https://github.com/ibevers))
+- 305 task update bioacoustic qc to run on b2ai dataset [#306](https://github.com/sensein/senselab/pull/306) ([@ibevers](https://github.com/ibevers))
+- Add 20 AQMs to bioacoustic qc [#304](https://github.com/sensein/senselab/pull/304) ([@ibevers](https://github.com/ibevers))
+- 269 task add storage of qc aqms and check results in a dataframe [#278](https://github.com/sensein/senselab/pull/278) ([@ibevers](https://github.com/ibevers))
+- 253 task implement initial version of bioacoustic quality control [#256](https://github.com/sensein/senselab/pull/256) ([@ibevers](https://github.com/ibevers))
+
+#### ⚠️ Pushed to `main`
+
+- Merge branch 'main' into 276-feature-implement-bioacoustic-quality-control-framework ([@ibevers](https://github.com/ibevers))
+
+#### Authors: 6
+
+- [@ibevers](https://github.com/ibevers)
+- Fabio Catania ([@fabiocat93](https://github.com/fabiocat93))
+- Rahul Brito ([@Rahul-Brito](https://github.com/Rahul-Brito))
+- rfbrito (rfbrito@node1702.inband)
+- rfbrito (rfbrito@node1703.inband)
+- Satrajit Ghosh ([@satra](https://github.com/satra))
+
+---
+
 # 1.1.0 (Fri Nov 07 2025)
 
 #### 🚀 Enhancement